did checkout,product page,cart,account settings and so on

+++ Backend/src/cart/cart.service.ts
@@ -1,12 +1,14 @@
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import { InjectRepository } from '@nestjs/typeorm';
+import { Repository } from 'typeorm';
 import { CreateCartDto } from './dto/create-cart.dto';
 import { UpdateCartDto } from './dto/update-cart.dto';
-import { InjectRepository } from '@nestjs/typeorm';
 import { Cart } from 'src/models/Cart.entity';
-import { Repository } from 'typeorm';
 import { Product } from 'src/models/Product.entity';
 import { User } from 'src/models/User.entity';
 import { CartItem } from 'src/models/CartItem.entity';
+import { UploadImagesService } from 'src/upload-images/upload-images.service';
+import { WishlistService } from 'src/wishlist/wishlist.service';
 
 @Injectable()
 export class CartService {
@@ -19,6 +21,8 @@ export class CartService {
     private readonly userRepository: Repository<User>,
     @InjectRepository(CartItem)
     private readonly cartItemRepository: Repository<CartItem>,
+    private readonly uploadImagesService: UploadImagesService,
+    private readonly wishlistService: WishlistService,
   ) {}
   async createCart(userId: string) {
     const existingCart = await this.cartRepository.findOne({
@@ -48,12 +52,30 @@ export class CartService {
     }
     const items = await this.cartItemRepository.find({
       where: { cart: { id: cart.id } },
-      relations: ['product'],
+      relations: ['product', 'product.category'],
     });
+    for (const item of items) {
+      const images = await this.uploadImagesService.imageRepository.find({
+        where: { entityType: 'product', entityId: item.product.id },
+      });
+      (item as any).images = images;
+    }
+    const subtotalPrice = items.reduce(
+      (acc, item) => acc + item.product.price * item.quantity,
+      0,
+    );
+    cart.subtotalPrice = subtotalPrice;
+    await this.cartRepository.save(cart);
     return { ...cart, items };
   }
   async addProductToCart(createCartDto: CreateCartDto) {
-    const { userId, productId, quantity } = createCartDto;
+    const {
+      userId,
+      productId,
+      quantity,
+      shouldIncrement = true,
+      attributes,
+    } = createCartDto;
     try {
       let cart = await this.getCart(userId);
       if (!cart) {
@@ -78,18 +100,32 @@ export class CartService {
           HttpStatus.BAD_REQUEST,
         );
       }
-      const existingItem = cart.items.find(
-        (item) => item.product?.id === product.id,
-      );
+      const existingItem = cart.items.find((item) => {
+        if (item.product.id !== product.id) return false;
+        return true;
+      });
+
       if (existingItem) {
-        existingItem.quantity += quantity;
+        existingItem.attributes.push(...attributes);
+        existingItem.quantity = shouldIncrement
+          ? existingItem.quantity + quantity
+          : quantity;
+        if (attributes?.length) {
+          const existingAttrSet = new Set(
+            existingItem.attributes.map((attr) => JSON.stringify(attr)),
+          );
+          const newAttributes = attributes.filter(
+            (attr) => !existingAttrSet.has(JSON.stringify(attr)),
+          );
+          existingItem.attributes.push(...newAttributes);
+        }
         await this.cartItemRepository.save(existingItem);
-        return this.cartRepository.save(cart);
       } else {
         const cartItem = this.cartItemRepository.create({
           cart: cart,
           product: product,
           quantity: quantity,
+          attributes: attributes,
         });
         await this.cartItemRepository.save(cartItem);
         cart.items.push(cartItem);
@@ -111,7 +147,7 @@ export class CartService {
     try {
       const user = await this.userRepository.findOne({ where: { id: userId } });
       if (!user) {
-        return new Error('User not found');
+        throw new HttpException('User not found', HttpStatus.NOT_FOUND);
       }
       let cart = await this.cartRepository.findOne({
         where: { user: { id: user.id } },
@@ -125,7 +161,7 @@ export class CartService {
         where: { id: productId },
       });
       if (!product) {
-        return new Error('Product not found');
+        throw new HttpException('Product not found', HttpStatus.NOT_FOUND);
       }
       if (quantity <= 0) {
         return this.removeItemFromCart(userId, productId);
@@ -139,7 +175,7 @@ export class CartService {
         }
         existingItem.quantity = quantity;
         await this.cartItemRepository.save(existingItem);
-        return this.cartRepository.save(cart);
+        return this.getCart(userId);
       } else {
         if (quantity < 0) {
           return this.removeItemFromCart(userId, productId);
@@ -147,7 +183,13 @@ export class CartService {
         return this.addProductToCart({ userId, productId, quantity });
       }
     } catch (error) {
-      throw new Error(error);
+      if (error instanceof HttpException) {
+        throw error;
+      }
+      throw new HttpException(
+        `Error updating cart item: ${error.message}`,
+        HttpStatus.INTERNAL_SERVER_ERROR,
+      );
     }
   }
 
@@ -178,16 +220,65 @@ export class CartService {
     return this.cartRepository.save(cart);
   }
 
-  async removeItemsCart(userId: string) {
-    const user = await this.userRepository.findOne({ where: { id: userId } });
-    const cart = await this.cartRepository.findOne({
-      where: { user: { id: user.id } },
-    });
-    if (!cart) {
-      return new HttpException('Cart not found', HttpStatus.NOT_FOUND);
+  async clearCart(userId: string) {
+    try {
+      const user = await this.userRepository.findOne({ where: { id: userId } });
+      if (!user) {
+        throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+      }
+      const cart = await this.cartRepository.findOne({
+        where: { user: { id: user.id } },
+        relations: ['items', 'items.product'],
+      });
+      console.log('cart', cart);
+      if (!cart) {
+        throw new HttpException('Cart not found', HttpStatus.NOT_FOUND);
+      }
+      if (!cart.items || cart.items.length === 0) {
+        return cart;
+      }
+      if (cart.items && cart.items.length > 0) {
+        await this.cartItemRepository.delete({ cart: { id: cart.id } });
+        cart.items = [];
+        cart.subtotalPrice = 0;
+        await this.cartRepository.save(cart);
+      }
+      return cart;
+    } catch (error) {
+      throw new HttpException(
+        error instanceof HttpException
+          ? error.getResponse()
+          : `Error clearing cart: ${error.message}`,
+        error instanceof HttpException
+          ? error.getStatus()
+          : HttpStatus.INTERNAL_SERVER_ERROR,
+      );
     }
-    await this.cartItemRepository.delete({ cart: { id: cart.id } });
-    return this.cartRepository.save(cart);
+  }
+  async removeAndAddToWishlist(userID: string, productID: string) {
+    const user = await this.userRepository.findOne({ where: { id: userID } });
+    if (!user) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+    const isAddedToWishlist = await this.wishlistService.findOne({
+      userID,
+      productID,
+    });
+    if (isAddedToWishlist.status == HttpStatus.OK)
+      throw new HttpException(
+        'Product already in wishlist',
+        HttpStatus.CONFLICT,
+      );
+    try {
+      this.removeItemFromCart(userID, productID);
+      await this.wishlistService.create({ userID, productID });
+      return {
+        message: 'Product removed from cart and added to wishlist',
+        status: HttpStatus.OK,
+      };
+    } catch (error) {
+      throw new HttpException(error.message, HttpStatus.INTERNAL_SERVER_ERROR);
+    }
   }
   async removeAll() {
     await this.cartRepository.query(`TRUNCATE carts RESTART IDENTITY CASCADE;`);
